Reorganize tracked object for imports
